pyproject.toml
[tool.poetry]
name = "examevoke"
version = "0.1.0"
description = "backend for examevoke"
authors = ["Isshiki <justjeevan07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
pre-commit = "^3.7.0"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88


[tool.black]
target_version = ['py38']